Ajuste nos Repository´s

+++ advocacy/src/main/java/github/io/advocacy/repository/ProcessRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProcessRepository extends JpaRepository<ProcessEntity, Integer> {
+public interface ProcessRepository extends JpaRepository<ProcessEntity, Long> {
 }
