+++ advocacy/src/main/java/github/io/advocacy/repository/SchedulingRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SchedulingRepository extends JpaRepository<SchedulingEntity, Integer> {
+public interface SchedulingRepository extends JpaRepository<SchedulingEntity, Long> {
 }
